Rename FlRenderer to FlCompositorOpenGL
