+++ engine/src/flutter/shell/platform/linux/BUILD.gn
@@ -105,6 +105,8 @@ source_set("flutter_linux_sources") {
     "fl_basic_message_channel.cc",
     "fl_binary_codec.cc",
     "fl_binary_messenger.cc",
+    "fl_compositor.cc",
+    "fl_compositor_opengl.cc",
     "fl_dart_project.cc",
     "fl_display_monitor.cc",
     "fl_engine.cc",
@@ -136,7 +138,6 @@ source_set("flutter_linux_sources") {
     "fl_plugin_registry.cc",
     "fl_pointer_manager.cc",
     "fl_renderable.cc",
-    "fl_renderer.cc",
     "fl_scrolling_manager.cc",
     "fl_settings.cc",
     "fl_settings_channel.cc",
@@ -213,6 +214,7 @@ executable("flutter_linux_unittests") {
     "fl_basic_message_channel_test.cc",
     "fl_binary_codec_test.cc",
     "fl_binary_messenger_test.cc",
+    "fl_compositor_opengl_test.cc",
     "fl_dart_project_test.cc",
     "fl_display_monitor_test.cc",
     "fl_engine_test.cc",
@@ -235,7 +237,6 @@ executable("flutter_linux_unittests") {
     "fl_platform_handler_test.cc",
     "fl_plugin_registrar_test.cc",
     "fl_pointer_manager_test.cc",
-    "fl_renderer_test.cc",
     "fl_scrolling_manager_test.cc",
     "fl_settings_handler_test.cc",
     "fl_settings_portal_test.cc",
