+++ engine/src/flutter/shell/platform/linux/fl_engine.cc
@@ -13,6 +13,7 @@
 #include "flutter/shell/platform/common/engine_switches.h"
 #include "flutter/shell/platform/embedder/embedder.h"
 #include "flutter/shell/platform/linux/fl_binary_messenger_private.h"
+#include "flutter/shell/platform/linux/fl_compositor_opengl.h"
 #include "flutter/shell/platform/linux/fl_dart_project_private.h"
 #include "flutter/shell/platform/linux/fl_display_monitor.h"
 #include "flutter/shell/platform/linux/fl_engine_private.h"
@@ -21,7 +22,6 @@
 #include "flutter/shell/platform/linux/fl_pixel_buffer_texture_private.h"
 #include "flutter/shell/platform/linux/fl_platform_handler.h"
 #include "flutter/shell/platform/linux/fl_plugin_registrar_private.h"
-#include "flutter/shell/platform/linux/fl_renderer.h"
 #include "flutter/shell/platform/linux/fl_settings_handler.h"
 #include "flutter/shell/platform/linux/fl_texture_gl_private.h"
 #include "flutter/shell/platform/linux/fl_texture_registrar_private.h"
@@ -49,7 +49,7 @@ struct _FlEngine {
   FlDisplayMonitor* display_monitor;
 
   // Renders the Flutter app.
-  FlRenderer* renderer;
+  FlCompositor* compositor;
 
   // Manages OpenGL contexts.
   FlOpenGLManager* opengl_manager;
@@ -252,26 +252,25 @@ static bool compositor_create_backing_store_callback(
     const FlutterBackingStoreConfig* config,
     FlutterBackingStore* backing_store_out,
     void* user_data) {
-  g_return_val_if_fail(FL_IS_RENDERER(user_data), false);
-  return fl_renderer_create_backing_store(FL_RENDERER(user_data), config,
-                                          backing_store_out);
+  FlEngine* self = static_cast<FlEngine*>(user_data);
+  return fl_compositor_create_backing_store(self->compositor, config,
+                                            backing_store_out);
 }
 
 // Called when the backing store is to be released.
 static bool compositor_collect_backing_store_callback(
     const FlutterBackingStore* backing_store,
     void* user_data) {
-  g_return_val_if_fail(FL_IS_RENDERER(user_data), false);
-  return fl_renderer_collect_backing_store(FL_RENDERER(user_data),
-                                           backing_store);
+  FlEngine* self = static_cast<FlEngine*>(user_data);
+  return fl_compositor_collect_backing_store(self->compositor, backing_store);
 }
 
 // Called when embedder should composite contents of each layer onto the screen.
 static bool compositor_present_view_callback(
     const FlutterPresentViewInfo* info) {
-  g_return_val_if_fail(FL_IS_RENDERER(info->user_data), false);
-  return fl_renderer_present_layers(FL_RENDERER(info->user_data), info->view_id,
-                                    info->layers, info->layers_count);
+  FlEngine* self = static_cast<FlEngine*>(info->user_data);
+  return fl_compositor_present_layers(self->compositor, info->view_id,
+                                      info->layers, info->layers_count);
 }
 
 // Flutter engine rendering callbacks.
@@ -484,7 +483,7 @@ static void fl_engine_dispose(GObject* object) {
 
   g_clear_object(&self->project);
   g_clear_object(&self->display_monitor);
-  g_clear_object(&self->renderer);
+  g_clear_object(&self->compositor);
   g_clear_object(&self->opengl_manager);
   g_clear_object(&self->texture_registrar);
   g_clear_object(&self->binary_messenger);
@@ -561,7 +560,7 @@ static FlEngine* fl_engine_new_full(FlDartProject* project,
   FlEngine* self = FL_ENGINE(g_object_new(fl_engine_get_type(), nullptr));
 
   self->project = FL_DART_PROJECT(g_object_ref(project));
-  self->renderer = fl_renderer_new(self);
+  self->compositor = FL_COMPOSITOR(fl_compositor_opengl_new(self));
   if (binary_messenger != nullptr) {
     self->binary_messenger =
         FL_BINARY_MESSENGER(g_object_ref(binary_messenger));
@@ -596,9 +595,9 @@ G_MODULE_EXPORT FlEngine* fl_engine_new_headless(FlDartProject* project) {
   return fl_engine_new(project);
 }
 
-FlRenderer* fl_engine_get_renderer(FlEngine* self) {
+FlCompositor* fl_engine_get_compositor(FlEngine* self) {
   g_return_val_if_fail(FL_IS_ENGINE(self), nullptr);
-  return self->renderer;
+  return self->compositor;
 }
 
 FlOpenGLManager* fl_engine_get_opengl_manager(FlEngine* self) {
@@ -668,7 +667,7 @@ gboolean fl_engine_start(FlEngine* self, GError** error) {
 
   FlutterCompositor compositor = {};
   compositor.struct_size = sizeof(FlutterCompositor);
-  compositor.user_data = self->renderer;
+  compositor.user_data = self;
   compositor.create_backing_store_callback =
       compositor_create_backing_store_callback;
   compositor.collect_backing_store_callback =
